+++ src/governance-compliance/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { EnterpriseRiskManagementDescriptor } from './enterprise-risk-management/enterprise-risk-management-descriptor.js';
 import { GovernanceOperatingModelDescriptor } from './operating-model/governance-operating-model-descriptor.js';
 import { GovernanceComplianceOverviewDescriptor } from './overview/governance-compliance-overview-descriptor.js';
 import { PolicyArchitectureDescriptor } from './policy-architecture/policy-architecture-descriptor.js';
@@ -11,6 +12,7 @@ export function addGovernanceCompliance(services) {
   services.registerSingleton('GovernanceComplianceOverviewDescriptor', () => new GovernanceComplianceOverviewDescriptor());
   services.registerSingleton('GovernanceOperatingModelDescriptor', () => new GovernanceOperatingModelDescriptor());
   services.registerSingleton('PolicyArchitectureDescriptor', () => new PolicyArchitectureDescriptor());
+  services.registerSingleton('EnterpriseRiskManagementDescriptor', () => new EnterpriseRiskManagementDescriptor());
 
   return services;
 }